configs: arm64: Make USB_DWC2 built-in
It has been decided that we should drop DWC_OTG support on the 64-bit
kernels. Its raison-d'etre on ARCH=arm is the FIQ FSM, which
significantly reduces the overheads of running the hardware in host
mode. Unfortunately, upstream Linux uses the DWC2 driver, which has had
a lot of attention over the years. In particular, there are a number of
situations where DWC_OTG fails where DWC2 works, despite the reduced
throughput.

In the ARMv8 kernel, FIQ support was missing, and is now just different
in an awkward way. It is possible to use DWC_OTG in the v8 kernel, but
only when the FIQ support is disabled, removing its advantage.

It therefore makes sense to transition to using the DWC2 driver in the
64-bit kernels, deprecating DWC_OTG. The first cautious step is to make
the DWC2 driver a built-in, so that either driver can be used for
booting. Unfortunately this increases the size of the kernel by ~200kB,
but the intention is later to demote DWC_OTG to a module or drop it
altogether.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>